# backend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
        "/health": {
            "get": {
                "tags": [
                    "health"
                ],
                "summary": "Health Check",
                "operationId": "health_check_health_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HealthCheckAPIResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/{email_address}": {
            "post": {
                "tags": [
                    "users"
                ],
                "summary": "Create User",
                "description": "Create a new user",
                "operationId": "create_user_users__email_address__post",
                "parameters": [
                    {
                        "name": "email_address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Email address of the user to create"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUser"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/me": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Get User",
                "description": "Get details of the currently authenticated user",
                "operationId": "get_user_users_me_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "users"
                ],
                "summary": "Delete User",
                "description": "Delete the currently authenticated user",
                "operationId": "delete_user_users_me_delete",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/users/{email_address}/login": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Login User",
                "description": "Login a user",
                "operationId": "login_user_users__email_address__login_get",
                "security": [
                    {
                        "HTTPBasic": []
                    }
                ],
                "parameters": [
                    {
                        "name": "email_address",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Email Address"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/users/logout": {
            "get": {
                "tags": [
                    "users"
                ],
                "summary": "Logout User",
                "description": "Logout the currently authenticated user",
                "operationId": "logout_user_users_logout_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/reviews": {
            "post": {
                "tags": [
                    "reviews"
                ],
                "summary": "Create Review",
                "description": "Create a new review",
                "operationId": "create_review_reviews_post",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateReview"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/reviews/{review_id}": {
            "put": {
                "tags": [
                    "reviews"
                ],
                "summary": "Update Review",
                "description": "Update an existing review",
                "operationId": "update_review_reviews__review_id__put",
                "parameters": [
                    {
                        "name": "review_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Review Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateReview"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "reviews"
                ],
                "summary": "Delete Review",
                "description": "Delete an existing review",
                "operationId": "delete_review_reviews__review_id__delete",
                "parameters": [
                    {
                        "name": "review_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Review Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CreateReview": {
                "properties": {
                    "stars": {
                        "type": "integer",
                        "title": "Stars"
                    },
                    "notes": {
                        "type": "string",
                        "title": "Notes"
                    },
                    "visit_again": {
                        "type": "boolean",
                        "title": "Visit Again"
                    },
                    "location_id": {
                        "type": "string",
                        "title": "Location Id"
                    }
                },
                "type": "object",
                "required": [
                    "stars",
                    "notes",
                    "visit_again",
                    "location_id"
                ],
                "title": "CreateReview"
            },
            "CreateUser": {
                "properties": {
                    "firstName": {
                        "type": "string",
                        "title": "Firstname"
                    },
                    "lastName": {
                        "type": "string",
                        "title": "Lastname"
                    },
                    "password": {
                        "type": "string",
                        "title": "Password"
                    }
                },
                "type": "object",
                "required": [
                    "firstName",
                    "lastName",
                    "password"
                ],
                "title": "CreateUser"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "HealthCheckAPIResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "title": "Status"
                    }
                },
                "type": "object",
                "required": [
                    "status"
                ],
                "title": "HealthCheckAPIResponse"
            },
            "UpdateReview": {
                "properties": {
                    "stars": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Stars"
                    },
                    "notes": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notes"
                    },
                    "visit_again": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Visit Again"
                    }
                },
                "type": "object",
                "title": "UpdateReview"
            },
            "User": {
                "properties": {
                    "emailAddress": {
                        "type": "string",
                        "format": "email",
                        "title": "Emailaddress"
                    },
                    "firstName": {
                        "type": "string",
                        "title": "Firstname"
                    },
                    "lastName": {
                        "type": "string",
                        "title": "Lastname"
                    }
                },
                "type": "object",
                "required": [
                    "emailAddress",
                    "firstName",
                    "lastName"
                ],
                "title": "User"
            },
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBasic": {
                "type": "http",
                "scheme": "basic"
            }
        }
    }
}